feat: added user not found error for delete user use case

+++ src/modules/users/infra/useCases/deleteUser/DeleteUserUseCase.ts
@@ -6,6 +6,12 @@ export class DeleteUserUseCase {
     ) {
     }
     async execute(id: string): Promise<void> {
+        const user = await this.usersRepository.findBy(id, 'id');
+
+        if (!user) {
+            throw new Error('User not found!');
+        }
+
         return await this.usersRepository.delete(id);
     }
 }
